% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{clear_CO}
\alias{clear_CO}
\title{Function to remove CO codes by frequency cutoff}
\usage{
clear_CO(CO, freq, freq_min = 1000)
}
\description{
Function to remove CO codes by frequency cutoff
}
\keyword{internal}
