% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{code_arrange}
\alias{code_arrange}
\title{Arrange codes}
\usage{
code_arrange(data)
}
\description{
let the "first col (V1) id" < "second col (V2) id"
and add the row with same V1 and V2
}
\keyword{internal}
