% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{concat_txt}
\alias{concat_txt}
\title{Function to concatentae text with string}
\usage{
concat_txt(vec1, vec2, concat = "->")
}
\description{
Function to concatentae text with string
}
\keyword{internal}
