% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{Evaluate_codi}
\alias{Evaluate_codi}
\title{Evaluate The Embedding (Codi Only)}
\usage{
Evaluate_codi(
  embed,
  AllRelationPairs,
  evatype = "all",
  prop = 0.3,
  normalize = TRUE,
  labels = NULL
)
}
\arguments{
\item{embed}{Embedding, should be normed, rownames are code names.}

\item{AllRelationPairs}{All relation pairs dataframe.}

\item{evatype}{Can be "train", "test", all the other character would be ignored.
\itemize{
\item{If \code{evatype="train"}, choose first \code{(prop * 100)}\% pairs as training pairs.}
\item{If \code{evatype="test"}, choose last \code{(prop * 100)}\% pairs as test pairs.}
\item{Otherwise, using all pairs to evaluate the embedding.}
}}

\item{prop}{There are 3 scenarios:}

\item{normalize}{\code{TRUE} or \code{TFALSE}, to normalize embedding before evaluation or not.
For evaluation purpose only, the final embedding result will NOT be normalized no matter what.}

\item{labels}{A character vector of labels in evaluations summary. Values can be 
\code{"PheCode-PheCode(sim)", "RXNORM-RXNORM(sim)", "LAB-LAB(sim)", 
"PheCode-PheCode(rela)", "PheCode-RXNORM(rela)", "PheCode-CCS(rela)", "PheCode-LAB(rela)"} .
If \code{NULL} then all labels will be used.}
}
\value{
A list.
}
\description{
Similar to function \code{Evaluate}, but for codi pairs only.
}
