% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{Evaluate_tmp}
\alias{Evaluate_tmp}
\title{Evaluate the embedding (codi & CUI)}
\usage{
Evaluate_tmp(embed, AllRelationPairs, evatype = "all", prop = 0.3)
}
\description{
Similar to function \code{Evaluate}, but just for codi pairs evaluation.
}
