% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed.R
\name{get_best_dim}
\alias{get_best_dim}
\title{Get Best Embedding from Summary}
\usage{
get_best_dim(summary, metric = "auc", by = "weighted.rela")
}
\arguments{
\item{summary}{the summary object from output of function \code{get_eval_embed}.}

\item{metric}{Metric for evaluation (Columns in summary evaluation dataframes). 
Options could be: #"pairs", "auc", "cut/0.01", "cut/0.05", "cut/0.1", "TPR/0.01", "TPR/0.05" or "TPR/0.1".
By default is auc.}

\item{by}{The evaluation object, options can be seems in plot summary file (the legend). By default is \code{weighted.rela}.}
}
\description{
Get Best Embedding from Summary
}
