% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{get_embed}
\alias{get_embed}
\title{Getting embedding from svd}
\usage{
get_embed(mysvd, d = 2000)
}
\arguments{
\item{mysvd}{The svd result.}

\item{d}{Dimension of the final embedding.}
}
\value{
A dataframe of embedding.
}
\description{
Getting embedding from svd
}
