% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embed_regression.R
\name{get_embed_regression}
\alias{get_embed_regression}
\title{Generate & Evaluate Embedding}
\usage{
get_embed_regression(
  embed_train,
  embed_valid,
  phecodes,
  dim,
  lambda_vec = c(seq(1, 51, 1) * 1e-06, seq(60, 1000, 50) * 1e-06),
  alpha = 0.25,
  custom_dict = NULL
)
}
\arguments{
\item{embed_train}{Tranning embedding for regression.}

\item{embed_valid}{Validation embedding for regression.}

\item{phecodes}{Interested phecodes.}

\item{dim}{Dimension used for embedding, for AIC/BIC calculation.}

\item{lambda_vec}{Lambda candidates for glmnet, it's very data specific. 
By default: \code{c(seq(1, 51, 1) \* 1e-6, seq(60, 1000, 50) \* 1e-6)}}

\item{alpha}{Alpha value for glmnet, by defaut is \code{0.25}}

\item{custom_dict}{Dictionary file for codes mapping. If not offered, the internal dictionary will be used.
Data structure:
\itemize{
\item{\code{code}}: codes like PheCode:714.1
\item{\code{desc}}: descriptions like rheumatoid arthritis
}}
}
\value{
A list of information including:
\itemize{
\item{\code{summary_data}}: Regression summary of selected codes, beta's, cosine values and code description.
\item{\code{Nlist}}: Number of non-zero beta's over lambda.
\item{\code{min_lambdas}}: The best lambda of mininmun AIC + Testing Residual for interested Phecodes.
\item{\code{eval_plots}}: Plots of Residuals over log(lambda) for interested Phecodes. 
\item{\code{wordcloud_plots}}: Word cloud plots for selected features magnified by cosine values. 
}
}
\description{
\code{get_embed_regression} acts as embedding regression to select related codes for interested PheCodes.
It returns a list of summary objects.
}
