% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed.R
\name{get_eval_embed}
\alias{get_eval_embed}
\title{Generate & Evaluate Embedding}
\usage{
get_eval_embed(
  CO_file,
  freq_file,
  dims = seq(100, 1000, 100),
  out_dir = NULL,
  freq_min = 1000,
  threshold = 10,
  normalize = TRUE,
  use.dataframe = FALSE,
  save.summary = TRUE
)
}
\arguments{
\item{CO_file}{Co-occurrence data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}. 
If \code{use.dataframe} = \code{TRUE}, then it should be a R dataframe variable.
The data should be a table with 3 columns index1, index2, count:
\itemize{
\item{\code{index1}}: Shows the index of code1.
\item{\code{index2}}: Shows the col index of code2.
\item{\code{count}}: Shows the counts for certain pair.
}
If the columns are not \code{index1}, \code{index2} and \code{count}, it will consider the first 3 columns as the corresponding columns.}

\item{freq_file}{Frequency count file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.
If \code{use.dataframe} = \code{TRUE}, then it should be a R dataframe variable.
The data should be a table with 4 columns index, code, description, freq_count:
\itemize{
\item{\code{index}}: Shows the index of code.
\item{\code{code}}: Shows the name of code.
\item{\code{description}}: Shows the description text of code.
\item{\code{freq_count}}: Shows the frequency count of code.
}
If the columns are not \code{index1}, \code{code}, \code{description} and \code{freq_count}, it will consider the first 4 columns as the corresponding columns.}

\item{dims}{A vector of numeric values for dimension, by default is \code{seq(100, 1000, 100)}.}

\item{out_dir}{Output folder, if \code{NULL} then by default set to your_working_directory/output.s}

\item{freq_min}{The frequency counts cutoff for code filtering. If the counts are less than \code{freq_min}, it’ll be filtered
out. By default is \code{1000}.}

\item{threshold}{Integer number, the threshold to get SPPMI matrix, by default is \code{10}.}

\item{normalize}{\code{TRUE} or \code{FALSE}, to normalize embedding or not. By default is \code{True}.}

\item{use.dataframe}{\code{TRUE} or \code{FALSE}. If \code{TURE}, \code{CO_file} and \code{freq_file} will be accepted as a R data frame variable, other than a file name.}

\item{save.summary}{\code{TRUE} or \code{FALSE}. If \code{FALSE}, the function will not save results to location \code{out_dir}.}
}
\value{
A list of information of meta-data, embedding & evaluation result. It will 
be saved in \code{out_dir} as \code{.Rdata} file.
}
\description{
\code{get_eval_embed} acts as embedding generation & evaluation from co-occurrence data file.
It returns a list of summary including meta-data, evaluation and embedding itself.
}
