% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Function to get plot data}
\usage{
get_plot_data(summary, col = "auc", dim_name = "dims", labels = NULL)
}
\description{
Function to get plot data
}
\keyword{internal}
