% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed.R
\name{get_report}
\alias{get_report}
\title{Generate Plot Report}
\usage{
get_report(
  summary,
  plot_val = "auc",
  knit_format = "html",
  labels = c("PheCode-PheCode(sim)", "RXNORM-RXNORM(sim)", "LAB-LAB(sim)",
    "PheCode-PheCode(rela)", "PheCode-RXNORM(rela)", "PheCode-LAB(rela)"),
  split_patterns = list(Similarity = "(sim)", Relation = "(rela)")
)
}
\arguments{
\item{summary}{A list of summary infomation generated by \code{get_eval_embed}.}

\item{plot_val}{Metric for evaluation (Columns in summary evaluation dataframes). 
Options could be: #"pairs", "auc", "cut/0.01", "cut/0.05", "cut/0.1", "TPR/0.01", "TPR/0.05" or "TPR/0.1".}

\item{knit_format}{Format of plot report, "pdf" or "html".}

\item{labels}{Labels to be used in plots, options are the row names of summary evaluation dataframes.}

\item{split_patterns}{Pattern to split lablels, normally not changed.}
}
\description{
\code{get_report} acts as plot report generation from summary returned by \code{get_eval_embed}.
It returns \code{NULL} but generates plots in \code{out_dir}, classified by similarity & relation.
Report format can be chosen in parameter \code{knit_format}.
}
