% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{getembedding}
\alias{getembedding}
\title{get embedding from SPPMI matrix}
\usage{
getembedding(SPPMI, d = 1500)
}
\arguments{
\item{SPPMI}{The SPPMI matrix obtained from the previous section.}

\item{d}{A numerical shows the target dimension of embedding.}
}
\description{
get embedding from SPPMI matrix
}
