% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{getSVD}
\alias{getSVD}
\title{Getting randomized svd from SPPI.}
\usage{
getSVD(SPPMI, dim)
}
\arguments{
\item{SPPMI}{The SPPMI matrix obtained from \code{getSPPMI}.}

\item{dim}{specifies the target rank of the low-rank decomposition used in rSVD, suggested to set as highest dimension + 1000.}
}
\value{
A list of matrix.
\itemize{
\item{\code{d} Array, singular values.}
\item{\code{u} Array, left singular vectors.}
\item{\code{v} Array, right singular vectors.}
}
}
\description{
Getting randomized svd from SPPI.
}
\keyword{internal}
