% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{path_chk}
\alias{path_chk}
\title{Function to check path}
\usage{
path_chk(path)
}
\arguments{
\item{path}{An full or relative path.}
}
\value{
A full path of input path.
}
\description{
Function to check path
}
\keyword{internal}
