% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{plt}
\alias{plt}
\title{Function to plot scatter-line by group}
\usage{
plt(data, x = "dims", y = "auc", group = "pairs", method = "plotly")
}
\description{
Function to plot scatter-line by group
}
\keyword{internal}
