% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{read_file}
\alias{read_file}
\title{Functions to read data file}
\usage{
read_file(file_name)
}
\arguments{
\item{file_name}{Data format can be .csv/.parquet/.Rdata.}
}
\value{
A DataFrame.
}
\description{
Functions to read data file
}
