% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{reshape_data}
\alias{reshape_data}
\title{Function to reshape plot data}
\usage{
reshape_data(data, x = "dims", y = "vals", group = "pairs")
}
\description{
Function to reshape plot data
}
\keyword{internal}
