% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{split_data}
\alias{split_data}
\title{Function to split plot data by group}
\usage{
split_data(
  plot_data,
  include.others = TRUE,
  dim_name = "dims",
  patterns = list(Similarity = "(sim)", Relation = "(rela)")
)
}
\description{
Function to split plot data by group
}
\keyword{internal}
