% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed.R
\name{get_eval_embed}
\alias{get_eval_embed}
\title{Generate & Evaluate Embedding}
\usage{
get_eval_embed(
  CO_file,
  dims = seq(100, 1000, 100),
  out_dir = NULL,
  CO_dict_file = NULL,
  freq_file = NULL,
  freq_min = 1000,
  threshold = 10,
  data_type = 1,
  HAM_file = NULL,
  ARP_file = NULL,
  normalize = TRUE,
  labels = NULL
)
}
\arguments{
\item{CO_file}{Co-occurrence data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.
The data should be a table with 3 columns V1, V2, V3:
\itemize{
\item{\code{V1}}: Shows the row id (code id) or row code pair.
\item{\code{V2}}: Shows the col id (code id) or row code pair..
\item{\code{V3}}: Shows the counts for certain pair.
}
Note: If V1 & V2 are code ids other than code pairs, a mapping dict offering information 
from id to code pair is needed. Please see \code{CO_dict_file} for details.}

\item{dims}{A vector of numeric values for dimension, by default is \code{seq(100, 1000, 100)}.}

\item{out_dir}{Output folder, if \code{NULL} then by default set to your_working_directory/output.}

\item{CO_dict_file}{A data file with two columns, where:
\itemize{
\item{\code{Column 1}: Shows the name of code pair.}
\item{\code{Column 2}: Shows the corresponding id number.}
}
File format can be \code{.csv}, \code{.parquet} or \code{.Rdata}.}

\item{freq_file}{A data file with two columns, where:
\itemize{
\item{\code{Column 1}: Shows the name of unique codes.}
\item{\code{Column 2}: Shows the corresponding frequency counts.}
}
Format of \code{freq_file} can be \code{.csv}, \code{.parquet} or \code{.Rdata}.}

\item{freq_min}{Frequency cutoff, codes with frequency count below \code{freq_min}
will be removed.}

\item{threshold}{Integer number, the threshold to get SPPMI matrix, by default is 10.}

\item{data_type}{If data Does not contain CUI codes, set as \code{1}. Otherwise, set as \code{2}.}

\item{HAM_file}{Multi-axial hierarchy data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.
By default it's set to \code{NULL} and read the default file within the package. 
If a file name is offered, it'll read it instead and replace the default file}

\item{ARP_file}{All relation pairs data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.}

\item{normalize}{\code{TRUE} or \code{TFALSE}, to normalize embedding or not.}

\item{labels}{A character vector of labels in evaluations summary. Values can be 
\code{"PheCode-PheCode(sim)", "RXNORM-RXNORM(sim)", "LAB-LAB(sim)", 
"PheCode-PheCode(rela)", "PheCode-RXNORM(rela)", "PheCode-CCS(rela)", "PheCode-LAB(rela)"} .
If \code{NULL} then all labels will be used.}
}
\value{
A list of information of meta-data, embedding & evaluation result. It will 
be saved in \code{out_dir} as \code{.Rdata} file.
}
\description{
\code{get_eval_embed} acts as embedding generation & evaluation from co-occurrence data file.
It returns a list of summary including meta-data, evaluation and embedding itself.
}
