% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed.R
\name{get_eval_embed}
\alias{get_eval_embed}
\title{Generate & Evaluate Embedding}
\usage{
get_eval_embed(
  CO_file,
  ARP_file,
  dims,
  CO_dict_file = NULL,
  out_dir = NULL,
  HAM_file = NULL,
  data_type = 1
)
}
\arguments{
\item{CO_file}{Co-ccurrence data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.
The data should be a table with 3 columns V1, V2, V3:
\itemize{
\item{\code{V1}}: Shows the row id (code id) or row code pair.
\item{\code{V2}}: Shows the col id (code id) or row code pair..
\item{\code{V3}}: Shows the counts for certain pair.
}
Note: If V1 & V2 are code ids other than code pairs, a mapping dict offering information 
from id to code pair is needed. Please see \code{CO_dict_file} for details.}

\item{ARP_file}{All relation pairs data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.}

\item{dims}{A vector of numeric values for dimension.}

\item{CO_dict_file}{A data file with two columns, where:
\itemize{
\item{\code{Column 1}: Shows the name of code pair.}
\item{\code{Column 2}: Shows the corresponding id number.}
}
File format can be \code{.csv}, \code{.parquet} or \code{.Rdata}.}

\item{out_dir}{Output folder, if \code{NULL} then by default set to your_working_directory/output.}

\item{HAM_file}{Multi-axial hierarchy data file with format \code{.csv}, \code{.parquet} or \code{.Rdata}.
By default it's set to \code{NULL} and read the default file within the package. 
If a file name is offered, it'll read it instead and replace the default file}

\item{data_type}{If data Does not contain CUI codes, set as \code{1}. Otherwise, set as \code{2}.}
}
\value{
A list of infomation of meta-data, embedding & evaluation result. It will 
be saved in \code{out_dir} as \code{.Rdata} file.
}
\description{
\code{get_eval_embed} acts as embedding generation & evaluation from co-occurrence data file.
It returns a list of summary including meta-data, evaluation and embedding itself.
}
