% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{getembedding}
\alias{getembedding}
\title{Get Embedding from SPPMI Matrix}
\usage{
getembedding(SPPMI, d = 1500)
}
\arguments{
\item{SPPMI}{The SPPMI matrix obtained from the previous section.}

\item{d}{A numerical shows the target dimension of embedding.}
}
\value{
A data matrix.
}
\description{
Get Embedding from SPPMI Matrix
}
