% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_embed_funcs.R
\name{getSPPMI}
\alias{getSPPMI}
\title{Get SPPMI matrix}
\usage{
getSPPMI(cooc, dict, code_LPcode)
}
\arguments{
\item{cooc}{Should be a triple form of cooc.}

\item{dict}{A data.frame.}

\item{code_LPcode}{The output of \code{get_rollup_dict(unique(c(data$V1, data$V2)))}}
}
\value{
A data matrix.
}
\description{
Obtain SPPMI Matrix from COOC Matrix in Triple Form.
}
